Vytvorenie dedenia pre Node
